Move cli to realtime_server package, add deploy script
