+++ deploy.sh
@@ -1,0 +1,31 @@
+#!/usr/bin/env bash
+set -e
+
+# Update
+sudo apt update
+
+# Install git
+sudo apt install -y git
+
+# Clone repo
+git clone https://github.com/getretake/retake.git
+cd retake
+git checkout cli
+
+# Install Docker and Docker Compose
+sudo apt-get install -y ca-certificates curl gnupg
+sudo install -m 0755 -d /etc/apt/keyrings
+curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+sudo chmod a+r /etc/apt/keyrings/docker.gpg
+echo \
+  "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
+  "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
+  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+sudo apt update
+sudo apt-get -y install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
+
+# Enable non-root docker
+sudo usermod -aG docker "${USER}" || true
+
+# Start stack
+sudo -E docker compose up -d
